requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
cryptography==42.0.5
debugpy==1.8.0
executing==2.0.1
frozenlist==1.4.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
gspread==6.0.2
httplib2==0.22.0
icecream==2.1.3
idna==3.6
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
magic-filter==1.0.12
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.5.8
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
pytz==2024.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
stegpy==0.0.1
StrEnum==0.4.15
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.12
yarl==1.9.4
